+++ application_project/app/src/main/java/com/alyak/detector/data/dto/map/KakaoPlaceResponse.kt
