+++ application_project/app/src/main/java/com/alyak/detector/ui/main/DoseStatusItem.kt
@@ -1,4 +1,4 @@
-package com.alyak.detector.ui.other
+package com.alyak.detector.ui.main
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -26,6 +26,7 @@ import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.alyak.detector.R
+import com.alyak.detector.ui.components.StatusBadge
 
 @Composable
 fun DoseStatusItem(
