+++ application_project/app/src/main/java/com/alyak/detector/ui/other/HospitalInfo.kt
@@ -31,6 +31,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.alyak.detector.R
+import com.alyak.detector.ui.components.StatusBadge
 
 @Composable
 fun HospitalInfo(
