+++ application_project/app/src/main/java/com/alyak/detector/ui/other/PillInfoBox.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.alyak.detector.R
 import com.alyak.detector.data.dto.pill.Pill
+import com.alyak.detector.ui.components.StatusBadge
 
 @Composable
 fun PillInfoBox(
