+++ application_project/app/src/main/java/com/alyak/detector/ui/signIn/SignInScreen.kt
@@ -27,7 +27,7 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.alyak.detector.R
-import com.alyak.detector.ui.other.FindPasswordForm
+import com.alyak.detector.ui.signIn.components.FindPasswordForm
 import com.alyak.detector.ui.signIn.state.ContentState
 import com.alyak.detector.ui.signUp.SignUpForm
 import com.alyak.detector.ui.signUp.SignUpViewModel
