+++ application_project/app/src/main/java/com/alyak/detector/ui/splash/SplashScreen.kt
